pyproject.toml

[project]
version = '0.0.0'
name = 'StandardBeams'

requires-python = '>=3.10'

dependencies = [
    'freecad-stubs>=1.0.21' ,
    'pyside6>=6.9.2'
]